+++ addons/main/magazines/CfgMagFAL.hpp
@@ -64,6 +64,28 @@ class hlc_30Rnd_762x51_TDim_fal: hlc_20Rnd_762x51_TDim_fal {
     mass = 22.2;
 };
 
+class hlc_30Rnd_762x51_B_L4: hlc_20Rnd_762x51_B_fal {
+    mass = 22.2;
+};
+class hlc_30rnd_762x51_MDIM_L4: hlc_20rnd_762x51_MDIM_FAL {
+    mass = 22.2;
+};
+class hlc_30rnd_762x51_M_L4: hlc_20rnd_762x51_M_FAL {
+    mass = 22.2;
+};
+class hlc_30Rnd_762x51_TDim_L4: hlc_20Rnd_762x51_TDim_fal {
+    mass = 22.2;
+};
+class hlc_30Rnd_762x51_mk316_L4: hlc_20Rnd_762x51_mk316_fal {
+    mass = 22.2;
+};
+class hlc_30Rnd_762x51_barrier_L4: hlc_20Rnd_762x51_barrier_fal {
+    mass = 22.2;
+};
+class hlc_30Rnd_762x51_T_L4: hlc_20Rnd_762x51_T_fal {
+    mass = 22.2;
+};
+
 class hlc_50Rnd_762x51_B_fal: 30Rnd_556x45_Stanag {
     mass = 30;
 };