+++ addons/main/script_component.hpp
@@ -1,9 +1,10 @@
 #define COMPONENT main
+#define COMPONENT_BEAUTIFIED NIArms
 #include "script_mod.hpp"
 
 // #define DEBUG_MODE_FULL
 // #define DISABLE_COMPILE_CACHE
-// #define CBA_DEBUG_SYNCHRONOUS
+// #define DEBUG_SYNCHRONOUS
 // #define ENABLE_PERFORMANCE_COUNTERS
 
 #ifdef DEBUG_ENABLED_WB_NIARMS
@@ -13,7 +14,5 @@
 #ifdef DEBUG_SETTINGS_WB_NIARMS
     #define DEBUG_SETTINGS DEBUG_SETTINGS_WB_NIARMS
 #endif
-
-#define MAGAZINE_WELL WB_Magazines
 
 #include "script_macros.hpp"