+++ addons/prep/script_component.hpp
@@ -1,0 +1,18 @@
+#define COMPONENT prep
+#define COMPONENT_BEAUTIFIED NIArms
+#include "..\main\script_mod.hpp"
+
+// #define DEBUG_MODE_FULL
+// #define DISABLE_COMPILE_CACHE
+// #define DEBUG_SYNCHRONOUS
+// #define ENABLE_PERFORMANCE_COUNTERS
+
+#ifdef DEBUG_ENABLED_WB_NIARMS_PREP
+    #define DEBUG_MODE_FULL
+#endif
+
+#ifdef DEBUG_SETTINGS_WB_NIARMS_PREP
+    #define DEBUG_SETTINGS DEBUG_SETTINGS_WB_NIARMS_PREP
+#endif
+
+#include "..\main\script_macros.hpp"